Move impls for Mask back into mask mod
